refactor(hydroflow_plus, paxos)!: simplify intervals and split Paxos-KV into separate module
