+++ hydroflow_plus_test/src/cluster/mod.rs
@@ -3,5 +3,6 @@ pub mod many_to_many;
 pub mod map_reduce;
 pub mod paxos;
 pub mod paxos_bench;
+pub mod paxos_kv;
 pub mod simple_cluster;
 pub mod two_pc;
